move right of reply views to own module
